pyproject.toml
[project]
name = "multi-turn-eval"
version = "0.1.0"
description = "Multi-turn LLM evaluation framework"
requires-python = ">=3.12"
dependencies = [
  "click>=8.3",
  "pipecat-ai[anthropic,aws,aws-nova-sonic,azure,cerebras,google,groq,openai,silero]==0.0.101",
  "google-genai>=1.56.0",
  "python-dotenv",
  "loguru",
  "sounddevice>=0.4.6",
  "soundfile>=0.12.1",
  "numpy>=1.26",
  "openai>=1.0.0",
  "claude-agent-sdk>=0.1.8",
  "torch>=2.9.1",
  "torchaudio>=2.9.1",
]

[project.scripts]
# Main CLI entry point: uv run multi-turn-eval <command>
multi-turn-eval = "multi_turn_eval.cli:main"
# Legacy utility scripts
record-turn-audio = "scripts.record_turn_audio:main"
resample-turn-audio = "scripts.resample_turn_audio:main"

[build-system]
requires = ["hatchling>=1.20"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
# Include both the new package structure and legacy scripts
packages = ["src/multi_turn_eval", "scripts", "benchmarks"]

[tool.hatch.build.targets.sdist]
only-include = ["src", "scripts", "benchmarks", "turns.py", "tools_schema.py", "system_instruction.py", "system_instruction_short.py"]

[tool.uv.sources]
# Use github branch of pipecat instead of PyPI (commented out to use PyPI versions)
# pipecat-ai = { git = "https://github.com/pipecat-ai/pipecat.git", rev = "main" }

# [tool.uv.sources]
# Use local checkout of pipecat instead of PyPI (commented out to use PyPI versions)
# "pipecat-ai" = { path = "vendor/pipecat" }
